refactor: create chatting room modal 컴포넌트 room type별로 분리
