+++ src/component/Chat/CreateChattingRoomModal/Private.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Private = () => {
+  return <div>Private</div>;
+};
