+++ src/component/Main/Main.tsx
@@ -14,7 +14,7 @@ import {
 import { ProfileModal } from '../Header/ProfileModal';
 import { ChatRoom } from '../Chat/ChatList/ChatRoom';
 import { GameMatch } from '../Game/GameMatch/GameMatch';
-import { CreateChattingRoomModal } from '../Chat/CreateChattingRoomModal';
+import { CreateChattingRoomModal } from '../Chat/CreateChattingRoomModal/CreateChattingRoomModal';
 import { AddFriendModal } from '../FriendList/AddFriendModal';
 import { FriendList } from '../FriendList/FriendList';
 import { FriendProfileModal } from '../FriendList/FriendProfileModal';
